requirements.txt
autopep8==1.5.7
greenlet==1.1.0
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
SQLAlchemy==1.4.21
toml==0.10.2
waitress==2.0.0
Werkzeug==2.0.1
